+++ simbad/lattice/tests/test_latticesearch.py
@@ -18,7 +18,7 @@ class Test(unittest.TestCase):
         lattice_db = simbad.LATTICE_DB
         cls.LS = LatticeSearch(lattice_db, os.getcwd())
     
-    def test_search(self):
+    def test_search_1(self):
         """Test case for LatticeSearch.search"""
         
         # Process the data from the toxd test case
@@ -129,7 +129,7 @@ class Test(unittest.TestCase):
         
         self.assertEqual(data, reference_data)
         
-    def test_calculate_niggli_cell(self):
+    def test_calculate_niggli_cell_1(self):
         """Test case for LatticeSearch.calculate_niggli_cell"""
         
         space_group = 'P212121'
@@ -140,7 +140,7 @@ class Test(unittest.TestCase):
         
         self.assertEqual(data, reference_data)
         
-    def test_check_sg(self):
+    def test_check_sg_1(self):
         """Test case for LatticeSearch.check_sg"""
         
         sg = 'A1'
@@ -148,6 +148,36 @@ class Test(unittest.TestCase):
         reference_data = 'P1'
         
         self.assertEqual(data, reference_data)
+
+    def test_pdb_in_results_1(self):
+        """Test case for LatticeSearch.pdb_in_results"""
+
+        from simbad.lattice.latticescore import LatticeSearchResult
+        results = []
+        score = LatticeSearchResult("4Z88", "", "", "62.420,62.492,163.64,90.234,90.000,119.962", 4.164, 3.892, 0.272, 23300, 0.1800)
+        results.append(score)
+
+        self.assertTrue(self.LS.pdb_in_results("4Z88", results)) 
+        
+    def test_pdb_in_results_2(self):
+        """Test case for LatticeSearch.pdb_in_results"""
+
+        from simbad.lattice.latticescore import LatticeSearchResult
+        results = []
+        score = LatticeSearchResult("4Z88", "", "", "62.420,62.492,163.64,90.234,90.000,119.962", 4.164, 3.892, 0.272, 23300, 0.1800)
+        results.append(score)
+
+        self.assertFalse(self.LS.pdb_in_results("4A23", results))
+ 
+    def test_pdb_in_results_3(self):
+        """Test case for LatticeSearch.pdb_in_results"""
+
+        from simbad.lattice.latticescore import LatticeSearchResult
+        results = []
+        score = LatticeSearchResult("4Z88", "", "", "62.420,62.492,163.64,90.234,90.000,119.962", 4.164, 3.892, 0.272, 23300, 0.1800)
+        results.append(score)
+
+        self.assertTrue(self.LS.pdb_in_results("4z88", results)) 
         
 if __name__ == "__main__":
     unittest.main()